% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapR.R
\name{bootstrapR}
\alias{bootstrapR}
\title{A bootstrap samples generator using R}
\usage{
bootstrapR(X, n)
}
\arguments{
\item{X}{the given set of one-dimension data}

\item{n}{the number of bootstraping}
}
\value{
a matrix of samples generated by bootstraping and the row of the sample matrix is \code{n}
}
\description{
A function for generating bootstrap samples of given data using R
}
\examples{
\dontrun{
X <- c(96,87,88,90,85)
outer <- bootstrapR(X, 10)
}
}
