% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrR.R
\name{corrR}
\alias{corrR}
\title{A correlation coefficient calculator using R}
\usage{
corrR(X, Y)
}
\arguments{
\item{X}{the first set of samples}

\item{Y}{the second set of samples with the same size of X}
}
\value{
the correlation coefficient of the data
}
\description{
A function for computing the correlation coefficient of two set of samples using R
}
\examples{
\dontrun{
X <- runif(100)
Y <- rnorm(100)
corr <- corrR(X,Y)
}
}
