% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covR.R
\name{covR}
\alias{covR}
\title{A covariance calculator using R}
\usage{
covR(X, Y)
}
\arguments{
\item{X}{the first set of samples}

\item{Y}{the second set of samples with the same size of X}
}
\value{
the covariance of the data
}
\description{
A function for computing the covaiance of two set of samples using R
}
\examples{
\dontrun{
X <- runif(100)
Y <- rnorm(100)
cov <- covR(X,Y)
}
}
