% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDER.R
\name{KDER}
\alias{KDER}
\title{A kernel density estimator using R}
\usage{
KDER(X, h, x)
}
\arguments{
\item{X}{the given set of data of one-dimension}

\item{h}{the bandwidth}

\item{x}{the estimated density function value where you want to know}
}
\value{
the estimated density function value
}
\description{
A function for estimating the density function of a given set of data using R
}
\examples{
\dontrun{
X <- rnorm(100)
x <- seq(-1, 1, by = 0.01)
h <- 0.1
for(i in 1:length(x)){
  y[i] <- KDER(X, h, x[i])
  }
plot(x, y, type="l")
}
}
