\name{SC19020-package}
\alias{SC19020-package}
\docType{package}
\title{
  This package is the final homework for Statistical Computing course consisting of two functions and all homework answers before.
}
\description{
  R version of bootstrap sampler and kernel density estimator are provided in this package (\code{bootstrapR} and \code{KDER}).
}
\author{
19020

Maintainer: 19020 <smyao@mail.ustc.edu.cn>
}
\keyword{ package }