% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwMetropolisC}
\alias{rwMetropolisC}
\title{Use three inputs to generate random walk metropolis samples using Rcpp.}
\usage{
rwMetropolisC(sigma, a, N)
}
\arguments{
\item{sigma}{the first predictor (numeirc)}

\item{a}{the second predictor (numeric)}

\item{N}{the third predictor (numeric)}
}
\value{
a random sample of size \code{N}
}
\description{
Use normal and uniform distribution to do this job
}
\examples{
\dontrun{
rw <- rwMetropolisC(2,25,1000)
}
}
