\name{SC19020-package}
\alias{SC19020-package}
\alias{SC19020}
\docType{package}
\title{
  This package provides several functions for for computing the covariance and correlation coefficient of two set of samples.
}
\description{
  R version of covariance and correlation coefficient function are provided in this package (\code{covR} and \code{corrR})
}
\author{
Yao Suming

Maintainer: Yao Suming <smyao@mail.ustc.edu.cn>
}
\keyword{ package }